Created the landing page, signup, login
Implemented the frontend for signup and login. After successful signup we are redirected to the login page and after successfull login we are redirected to the welcome page that shows our name and we can press the stop button to come back to the landing page. The Navbar is not modified yet.